+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Make it possible to configure arrow writer when writing batch records to parquet.
+- Lower peak memory while building a `Graph` by ~65%.
 
 ### Deprecated